% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_name_wt.R
\name{adj_name_wt}
\alias{adj_name_wt}
\title{Adjust probabilistic matching weights.}
\usage{
adj_name_wt(data, wt = 0.9)
}
\arguments{
\item{data}{a dataframe created by matching two datasets}

\item{wt}{the weight affected matches will be assigned if greater than the original weight.}
}
\description{
Adjust the weight of matching in instances where one dataset has a multi-part name and the other doesn't. This can be used to make up for the relative under weighting of these matches.
}
