% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_logical.R
\name{catch_logical}
\alias{catch_logical}
\title{Catch logical values.}
\usage{
catch_logical(x)
}
\arguments{
\item{x}{a dataframe}
}
\description{
Columns with many missing values may be read in as logical, losing data. This function will catch any unnoticed instances of this.
}
