% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_missing.R
\name{clean_missing}
\alias{clean_missing}
\title{Generate age groups.}
\usage{
clean_missing(x, nval = 9, cval = "")
}
\arguments{
\item{x}{variable}

\item{nval}{cleaned value for numeric vars}

\item{cval}{cleaned value for text vars}
}
\description{
Generate age groups based on 'age' variable, up to 55+. Commonly used in HC reports.
}
