% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_notification.R
\name{email_notification}
\alias{email_notification}
\title{Send email notifications}
\usage{
email_notification(address, subject, body)
}
\arguments{
\item{address}{either a single email address or a vector of multiple addresses.}

\item{subject}{email subject text}

\item{body}{email body text}
}
\description{
Send emails through the Hennepin County mail API, dependant on HCPH Key Vault access.
}
