% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{data}
\name{health_fs}
\alias{health_fs}
\title{Data lake volume path.}
\format{
An object of class \code{character} of length 1.
}
\usage{
health_fs
}
\description{
Beginning of all volume paths
}
\keyword{datasets}
