% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mpls_nbhds}
\alias{mpls_nbhds}
\title{Minneapolis neighborhood shapefiles}
\format{
Not going to document everything here. Contains a geometry column:
\describe{
  \item{id}{integer from 1 to 5}
  \item{names}{letter, from A to E}
  ...
}
}
\usage{
mpls_nbhds
}
\description{
Shapefiles downloaded from Minneapolis open data. Last updated on 02/24.
}
\keyword{datasets}
