% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_summarise.R
\name{multi_summarise}
\alias{multi_summarise}
\title{Summarise across multiple variables}
\usage{
multi_summarise(data, group_cols, sum_cols)
}
\arguments{
\item{data}{dataframe}

\item{group_cols}{columns used for grouping}

\item{sum_cols}{columns used for summarizing}
}
\description{
Summarise across multiple variables
}
