% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv_nologic.R
\name{read_csv_nologic}
\alias{read_csv_nologic}
\title{Read csv files.}
\usage{
read_csv_nologic(x, ...)
}
\arguments{
\item{x}{filepath}
}
\description{
Read csv files, and convert all columns that would be read in as logical as character. This prevents losing information in columns that are mostly blank.
}
