% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_delim_nologic.R
\name{read_delim_nologic}
\alias{read_delim_nologic}
\title{Read delimited files.}
\usage{
read_delim_nologic(x, ...)
}
\arguments{
\item{x}{filepath}
}
\description{
Read delimited files, and convert all columns that would be read in as logical as character. This prevents losing information in columns that are mostly blank.
}
