% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xlsx_nologic.R
\name{read_xlsx_nologic}
\alias{read_xlsx_nologic}
\title{Read xlsx files.}
\usage{
read_xlsx_nologic(x, ...)
}
\arguments{
\item{x}{filepath}
}
\description{
Read xlsx files, and convert all columns that would be read in as logical as character. This prevents losing information in columns that are mostly blank.
}
